Fix transition logging (#3303)
